+++ drivers/dac/dac_demo/dac_demo.c
@@ -191,7 +191,7 @@ int32_t dac_write_samples(void* dev, uint16_t* buff, uint32_t samples)
  * @param attr_id - Attribute ID
  * @return Length of chars written in buf, or negative value on failure.
  */
-int get_dac_demo_attr(void *device, char *buf, size_t len,
+int get_dac_demo_attr(void *device, char *buf, uint32_t len,
 			  const struct iio_ch_info *channel, intptr_t attr_id)
 {
 	struct dac_demo_desc *desc;
@@ -222,7 +222,7 @@ int get_dac_demo_attr(void *device, char *buf, size_t len,
  * @param attr_id - Attribute ID
  * @return: Number of bytes written to device, or negative value on failure.
  */
-int set_dac_demo_attr(void *device, char *buf, size_t len,
+int set_dac_demo_attr(void *device, char *buf, uint32_t len,
 			  const struct iio_ch_info *channel, intptr_t attr_id)
 {
 	struct dac_demo_desc *desc;
