+++ drivers/photo-electronic/adpd188/iio_adpd188.c
@@ -48,16 +48,16 @@
 #include "no-os/util.h"
 
 static int adpd188_iio_read_offset_chan(void *device, char *buf,
-		size_t len,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv);
 static int adpd188_iio_change_offset_chan(void *device, char *buf,
-		size_t len,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv);
-static int adpd188_iio_read_raw_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_read_raw_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv);
-static int adpd188_iio_read_odr_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_read_odr_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv);
-static int adpd188_iio_change_odr_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_change_odr_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv);
 
 static struct iio_attribute adpd188_channel_attributes[] = {
@@ -125,7 +125,7 @@ static struct iio_channel adpd188_channels[] = {
  * @return Length of the response or 0 in case of error.
  */
 static int adpd188_iio_read_offset_chan(void *device, char *buf,
-		size_t len,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
@@ -153,7 +153,7 @@ static int adpd188_iio_read_offset_chan(void *device, char *buf,
  * @return Length of the input buffer.
  */
 static int adpd188_iio_change_offset_chan(void *device, char *buf,
-		size_t len,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
@@ -190,7 +190,7 @@ static int adpd188_iio_change_offset_chan(void *device, char *buf,
  * @param priv - Private pointer not used in this case.
  * @return Length of the response or 0 in case of error.
  */
-static int adpd188_iio_read_odr_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_read_odr_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
@@ -214,7 +214,7 @@ static int adpd188_iio_read_odr_chan(void *device, char *buf, size_t len,
  * @param priv - Private pointer not used in this case.
  * @return Length of the output buffer.
  */
-static int adpd188_iio_change_odr_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_change_odr_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
@@ -278,7 +278,7 @@ static int32_t adpd188_iio_standby_mode(struct adpd188_dev *desc)
  * @param priv - Private pointer not used in this case.
  * @return Length of the response or 0 in case of error.
  */
-static int adpd188_iio_read_raw_chan(void *device, char *buf, size_t len,
+static int adpd188_iio_read_raw_chan(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
